added a socket and a member function
added a connection socket that displays the users name and what room it joined when a connection has occured, also added a function in database that adds a member when a user has joined.

an error also occured when you used the functions in the user class: sqlite3.ProgrammingError: SQLite objects created in a thread can only be used in that same thread. The object was created in thread id 140609413226496 and this is thread id 140608974878272.

the error is quickly handled by discarding the __init__() functions members and just adding them in every function